extract region menu
